requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.112.0
fastjsonschema==2.20.0
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyterlab==4.2.4
nltk==3.8.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
notebook==7.2.1
numpy==1.25.0
openai==1.39.0
packaging==24.1
pillow==10.4.0
pinecone==5.0.0
pydantic==2.8.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.3.0
streamlit==1.37.1
threadpoolctl==3.5.0
torch==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
